requirements.txt
pillow
pdf2docx
pandas
pdf2image